Extract commonalities from Address,Email and Phone class and  put them into a super class Contact
